third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/no-style-sharing-with-distributed-nodes-expected.txt

CONSOLE ERROR: line 6: Uncaught TypeError: document.querySelector(...).createShadowRoot is not a function
